resources: add a script to disable systemd services for arm isa

+++ src/npb-24.04-imgs/scripts/post-installation.sh
@@ -16,9 +16,14 @@ make clean
 make suite M5_ANNOTATION=1
 echo "Disabling network by default"
 echo "See README.md for instructions on how to enable network"
-mv /etc/netplan/50-cloud-init.yaml /etc/netplan/50-cloud-init.yaml.bak
+
+if [ -f /etc/netplan/50-cloud-init.yaml ]; then
+    mv /etc/netplan/50-cloud-init.yaml /etc/netplan/50-cloud-init.yaml.bak
+elif [ -f /etc/netplan/00-installer-config.yaml ]; then
+    mv /etc/netplan/00-installer-config.yaml /etc/netplan/00-installer-config.yaml.bak
+    netplan apply
+fi
+
 # Disable systemd service that waits for network to be online
 systemctl disable systemd-networkd-wait-online.service
 systemctl mask systemd-networkd-wait-online.service
-
-netplan apply
